@@ -18,10 +18,10 @@
         "issues": "https://github.com/oscarotero/folk/issues"
     },
     "require": {
-        "php": ">=5.5.0",
+        "php": "^7.0",
         "zendframework/zend-diactoros": "^1.3",
         "league/plates": "^3.1",
-        "fol/core": "^11.1",
+        "oscarotero/fol": "^1.0",
         "form-manager/form-manager": "^5.0.4",
         "gettext/gettext": "^4.0",
         "oscarotero/inline-svg": "^2.0",
@@ -43,7 +43,9 @@
         }
     },
     "autoload-dev": {
-        "psr-4": { "Demo\\": "demo/" }
+        "psr-4": {
+            "Demo\\": "demo/"
+        }
     },
     "require-dev": {
         "symfony/yaml": "^3.0",
