@@ -33,7 +33,7 @@
 			<?= $entity->title ?>
 		</a>
 
-		<input id="search-entity" name="query" type="search" placeholder="<?= isset($placeholder) ? $placeholder : p__('search', 'Search %s...', strtolower($entity->title)) ?>" value="<?= isset($search) ? $search->buildQuery() : ''; ?>">
+		<input id="search-entity" name="query" type="search" placeholder="<?= $placeholder ?? p__('search', 'Search %s...', strtolower($entity->title)) ?>" value="<?= isset($search) ? $search->buildQuery() : ''; ?>">
 		<input type="hidden" name="page" value="1">
 		<button type="submit">
 			<?= $this->icon('action/search'); ?>
