@@ -3,17 +3,17 @@
 namespace Demo\Entities;
 
 use Folk\Entities\Yaml;
-use FormManager\Builder;
-use FormManager\InvalidValueException;
+use Folk\Formats\Group;
+use Folk\Formats\FormatFactory;
 
 class Tags extends Yaml
 {
-    protected function getBasePath()
+    protected function getBasePath(): string
     {
         return __DIR__.'/json';
     }
 
-    public function getScheme(Builder $builder)
+    public function getScheme(FormatFactory $builder): Group
     {
         return $builder->group([
             'name' => $builder->text()
