@@ -10,7 +10,7 @@ use Middlewares\Utils\Factory;
 
 class UpdateEntityField extends Entity
 {
-    public function json(Request $request, Admin $app, $entityName)
+    public function json(Request $request, Admin $app, string $entityName)
     {
         $id = $request->getAttribute('id');
         $field = $request->getAttribute('field');
