@@ -2,15 +2,14 @@
 
 namespace Folk\Formats;
 
-use FormManager\Builder;
 use Folk\Entities\EntityInterface;
 use Folk\SearchQuery;
 
 class RelationMany extends RelationOne implements FormatInterface
 {
-    public function __construct(Builder $builder, EntityInterface $related, SearchQuery $search = null)
+    public function __construct(FormatFactory $factory, EntityInterface $related, SearchQuery $search = null)
     {
-        parent::__construct($builder, $related, $search);
+        parent::__construct($factory, $related, $search);
 
         $this->input->multiple();
     }
